+++ src/main/resources/liquidbounce.mixins.json
@@ -27,6 +27,7 @@
     "minecraft.client.MixinClientLevel",
     "minecraft.client.MixinGameNarrator",
     "minecraft.client.MixinGameType",
+    "minecraft.client.MixinGui",
     "minecraft.client.MixinKeyboardHandler",
     "minecraft.client.MixinKeyboardInput",
     "minecraft.client.MixinKeyMapping",
@@ -70,12 +71,12 @@
     "minecraft.gui.MixinChatScreen",
     "minecraft.gui.MixinChatScreenAccessor",
     "minecraft.gui.MixinCommandSuggestions",
-    "minecraft.gui.MixinHudAccessor",
-    "minecraft.gui.MixinHud",
     "minecraft.gui.MixinGuiGraphicsExtractor",
     "minecraft.gui.MixinGuiRenderer",
     "minecraft.gui.MixinGuiRenderState",
     "minecraft.gui.MixinGuiRenderStateLayer",
+    "minecraft.gui.MixinHud",
+    "minecraft.gui.MixinHudAccessor",
     "minecraft.gui.MixinLoadingOverlay",
     "minecraft.gui.MixinLocatorBar",
     "minecraft.gui.MixinPlayerTabOverlay",
@@ -157,10 +158,10 @@
     "minecraft.world.chuck.MixinSkyLightEngine",
     "sodium.MixinSodiumAbstractBlockRenderContext",
     "sodium.MixinSodiumLightDataAccess",
+    "truffle.MixinCallerSensitiveDetection",
     "truffle.MixinCommonJSRequireBuiltin",
     "truffle.MixinNpmCompatibleESModuleLoader",
     "truffle.MixinTruffleLanguage",
-    "truffle.MixinCallerSensitiveDetection",
     "viaversion.MixinPacketWrapper"
   ],
   "server": [],
